test: quarantine foreign widget renderer coverage
